Refactor QuarterlyList composables
