+++ features/lessons/src/main/java/com/cryart/sabbathschool/lessons/ui/quarterlies/model/GroupedQuarterlies.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021. Adventech <info@adventech.io>
+ * Copyright (c) 2022. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package com.cryart.sabbathschool.lessons.ui.quarterlies.components
+package com.cryart.sabbathschool.lessons.ui.quarterlies.model
 
 import androidx.compose.runtime.Immutable
 import app.ss.models.QuarterlyGroup
