+++ features/lessons/src/test/java/com/cryart/sabbathschool/lessons/ui/quarterlies/QuarterliesViewModelTest.kt
@@ -31,8 +31,8 @@ import app.ss.models.auth.SSUser
 import com.cryart.sabbathschool.core.extensions.prefs.SSPrefs
 import com.cryart.sabbathschool.core.misc.SSConstants
 import com.cryart.sabbathschool.core.response.Resource
-import com.cryart.sabbathschool.lessons.ui.quarterlies.components.GroupedQuarterlies
-import com.cryart.sabbathschool.lessons.ui.quarterlies.components.placeHolderQuarterlies
+import com.cryart.sabbathschool.lessons.ui.quarterlies.model.GroupedQuarterlies
+import com.cryart.sabbathschool.lessons.ui.quarterlies.model.placeHolderQuarterlies
 import com.cryart.sabbathschool.test.coroutines.TestDispatcherProvider
 import io.mockk.coEvery
 import io.mockk.every
